mc-9141 Update models interfaces to allow creation of new ModelVersions/Branches
authentication tests
